+++ public/app/features/dashboard/services/DashboardSrv.ts
@@ -8,7 +8,6 @@ import { GrafanaRootScope } from 'app/routes/GrafanaCtrl';
 import { BackendSrv } from 'app/core/services/backend_srv';
 import { ILocationService } from 'angular';
 import { AppEvents } from '@grafana/data';
-import { PanelEvents } from '@grafana/data';
 
 interface DashboardSaveOptions {
   folderId?: number;
@@ -27,7 +26,7 @@ export class DashboardSrv {
     private $location: ILocationService
   ) {
     appEvents.on(CoreEvents.saveDashboard, this.saveDashboard.bind(this), $rootScope);
-    appEvents.on(PanelEvents.panelChangeView, this.onPanelChangeView);
+    appEvents.on(CoreEvents.panelChangeView, this.onPanelChangeView);
     appEvents.on(CoreEvents.removePanel, this.onRemovePanel);
   }
